Add menu for switching labels
The `react-aria-menubutton` package has a number of a11y features which is why I choose it.

In order to wire up the label changing menu, either all LabelInfo objects will need to be passed into each label button, or we could switch to redux.

I'm leaning towards switching to redux.